+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## Breaking changes
 
 * `useStrict(boolean)` was dropped, use `configure({enforceActions: boolean})` instead
+* `isolateGlobalState` was dropped, use `configure({ isolateGlobalState: true})` instead
 * `expr` is moved to mobx-utils. Remember, `expr(fn)` is just sugar for `computed(fn).get()`
 * `createTransformer` is moved to mobx-utils
 * Passing `context` explicitly to `autorun`, `reaction` etc is no longer supported. Use arrow functions or function.bind instead.